+++ src-docs/src/views/xy_chart_histogram/histogram_example.js
@@ -12,8 +12,8 @@ import {
   EuiCode,
   EuiCallOut,
   EuiLink,
-  EuiHistogramSeries,
 } from '../../../../src/components';
+import { EuiHistogramSeries } from '../../../../src/experimental';
 
 export const XYChartHistogramExample = {
   title: 'Histograms',