+++ src/components/tool_tip/icon_tip.js
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import { EuiIcon } from '../icon';
 import { EuiToolTip } from './tool_tip';
 
-export const EuiIconTip = ({ type, 'aria-label': ariaLabel, color, ...rest }) => (
+export const EuiIconTip = ({ type, 'aria-label': ariaLabel, color, size, ...rest }) => (
   <EuiToolTip {...rest}>
-    <EuiIcon tabIndex="0" type={type} color={color} aria-label={ariaLabel} />
+    <EuiIcon tabIndex="0" type={type} color={color} size={size} aria-label={ariaLabel} />
   </EuiToolTip>
 );
 
@@ -20,6 +20,11 @@ EuiIconTip.propTypes = {
    * The icon color.
    */
   color: PropTypes.string,
+
+  /**
+   * The icon size.
+   */
+  size: PropTypes.string,
 
   /**
    * Explain what this icon means for screen readers.