+++ src/index.js
@@ -1,3 +1,4 @@
 export * from './components';
+export * from './experimental';
 export * from './services';
 export * from './utils';